+++ FirstProjectCardGame/Model/Card.m
@@ -15,9 +15,9 @@
 
 @implementation Card
 
--(int)match:(NSArray *)otherCards
+-(NSInteger)match:(NSArray *)otherCards
 {
-    int score = 0;
+    NSInteger score = 0;
     
     for(Card *card in otherCards){
